requirements.txt
langchain
python-dotenv
langgraph
langchain-groq
psycopg2
mysql-connector-python
pandas
matplotlib